refactoring project for IntelliJ and Gradle
