+++ IntelliJ/src/main/java/com/nuix/javaenginesimple/LicenseFeaturesLogger.java
