+++ IntelliJ/src/main/java/com/nuix/javaenginesimple/LicenseFilter.java
