+++ IntelliJ/src/main/java/com/nuix/javaenginesimple/NuixDiagnostics.java
