+++ IntelliJ/src/main/java/com/nuix/javaenginesimple/ThirdPartyDependencyChecker.java
