+++ IntelliJ/src/main/java/com/nuix/javaenginesimple/examples/RevealDecryptionPasswordExample.java
@@ -2,7 +2,9 @@ package com.nuix.javaenginesimple.examples;
 
 import java.io.File;
 import java.io.IOException;
+import java.util.Set;
 import java.util.function.Consumer;
+import java.util.function.Function;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -13,22 +15,22 @@ import com.nuix.javaenginesimple.LicenseFilter;
 import com.nuix.javaenginesimple.NuixDiagnostics;
 
 import nuix.Case;
+import nuix.Item;
 import nuix.Utilities;
 
-/***
- * An example demonstrating opening an existing Nuix case.
- * 
- * See BasicInitializationExample for more details regarding the basic initialization steps being taken.
- * @author Jason Wells
- *
- */
-public class OpenCaseExample {
+public class RevealDecryptionPasswordExample {
 	private static Logger logger = null;
 
 	public static void main(String[] args) throws Exception {
-		String logDirectory = String.format("C:\\NuixEngineLogs\\%s",DateTime.now().toString("YYYYMMDD_HHmmss"));
-		EngineWrapper wrapper = new EngineWrapper("D:\\engine-releases\\9.2.4.392",logDirectory);
-		logger = LogManager.getLogger(LoadDataIntoCaseExample.class);
+		// Specify a custom location for our log files
+		String logDirectory = String.format("%s/%s",System.getProperty("nuix.logDir"),DateTime.now().toString("YYYYMMDD_HHmmss"));
+
+		// Create an instance of engine wrapper, which will do the work of getting the Nuix bits initialized.
+		// Engine wrapper will need to know what directory your engine release resides.
+		EngineWrapper wrapper = new EngineWrapper(System.getProperty("nuix.engineDir"), logDirectory);
+
+		// Relying on log4j2 initializations in EngineWrapper creation, so we wait until after that to fetch our logger
+		logger = LogManager.getLogger(RevealDecryptionPasswordExample.class);
 		
 		LicenseFilter licenseFilter = wrapper.getLicenseFilter();
 		licenseFilter.setMinWorkers(4);
@@ -45,6 +47,15 @@ public class OpenCaseExample {
 			logger.info("License password was provided via argument -DLicense.Password");
 		}
 		
+		Function<char[],String> passwordHandler = new Function<char[],String>(){
+			@Override
+			public String apply(char[] t) {
+				// We convert the character array containing the password to
+				// a String and then return that to the outside
+				return new String(t);
+			}
+		};
+		
 		try {
 			wrapper.trustAllCertificates();
 			wrapper.withCloudLicense(licenseUserName, licensePassword, new Consumer<Utilities>() {
@@ -57,10 +68,13 @@ public class OpenCaseExample {
 						logger.info(String.format("Opening case: %s",caseDirectory.toString()));
 						nuixCase = utilities.getCaseFactory().open(caseDirectory);
 						logger.info("Case opened");
-						
-						// *** Do things with the case here ***
-						long emailCount = nuixCase.count("kind:email");
-						logger.info(String.format("Email Item Count: %s",emailCount));
+
+						Set<Item> items = nuixCase.searchUnsorted("flag:encrypted");
+						for(Item item : items) {
+							String guid = item.getGuid();
+							String password = item.revealDecryptionPassword(passwordHandler);
+							logger.info(String.format("Item with GUID %s was decrypted with password: %s", guid, password));
+						}
 						
 						// Note that nuixCase is closed in finally block below
 					} catch (IOException exc) {
@@ -83,4 +97,3 @@ public class OpenCaseExample {
 		}
 	}
 }
-
