--- Java/src/main/resources/log4j.properties
@@ -1,21 +1,0 @@
-# Default log level is INFO, change to DEBUG to get more information.
-nuix.loglevel=INFO
-
-# Default log level for Console.  We use this to turn on logging in Jenkins.
-nuix.console.loglevel=INFO
-log4j.rootLogger=${nuix.loglevel}, NUIX_LOG, NUIX
-
-# Writes normal logging to nuix.log in an appropriate location.
-log4j.appender.NUIX_LOG=org.apache.log4j.rolling.RollingFileAppender
-log4j.appender.NUIX_LOG.file=nuix.log
-log4j.appender.NUIX_LOG.encoding=UTF-8
-log4j.appender.NUIX_LOG.append=true
-log4j.appender.NUIX_LOG.layout=com.nuix.common.log.NuixLayout
-log4j.appender.NUIX_LOG.rollingPolicy=com.nuix.common.log.NuixRollingPolicy
-log4j.appender.NUIX_LOG.rollingPolicy.fileNamePattern=nuix.%d.log
-
-log4j.appender.NUIX=org.apache.log4j.ConsoleAppender
-log4j.appender.NUIX.Target=System.out
-log4j.appender.NUIX.layout=org.apache.log4j.PatternLayout
-log4j.appender.NUIX.layout.ConversionPattern=%d{yyyy-MM-dd HH:mm:ss} %-5p %c:%M():%L - %m%n
-log4j.logger.com.nuix.engine=DEBUG,NUIX
